Rename `typing.py` to `utils.py` since `StagePositionTuple` affects more than typing
