+++ src/instamatic/microscope/interface/fei_microscope.py
@@ -11,7 +11,7 @@ from numpy import pi
 from instamatic import config
 from instamatic.exceptions import FEIValueError, TEMCommunicationError
 from instamatic.microscope.base import MicroscopeBase
-from instamatic.microscope.typing import StagePositionTuple
+from instamatic.microscope.utils import StagePositionTuple
 from instamatic.typing import float_deg, int_nm
 
 logger = logging.getLogger(__name__)
