+++ src/instamatic/microscope/utils.py
